Do some light reorganization
