+++ interp/src/primitives/mod.rs
@@ -1,8 +1,8 @@
-mod primitive;
-pub use primitive::Entry;
-pub use primitive::Named;
-pub use primitive::Primitive;
-pub use primitive::Serializable;
+mod primitive_traits;
+pub use primitive_traits::Entry;
+pub use primitive_traits::Named;
+pub use primitive_traits::Primitive;
+pub use primitive_traits::Serializable;
 
 pub mod combinational;
 pub(super) mod prim_utils;
