+++ interp/src/structures/environment.rs
@@ -224,7 +224,7 @@ impl InterpreterState {
             "std_wire" => {
                 Box::new(combinational::StdWire::new(params, cell_qin))
             }
-            // Unsigned Comparsion
+            // Unsigned Comparison
             "std_ge" => Box::new(combinational::StdGe::new(params, cell_qin)),
             "std_le" => Box::new(combinational::StdLe::new(params, cell_qin)),
             "std_lt" => Box::new(combinational::StdLt::new(params, cell_qin)),
@@ -257,20 +257,20 @@ impl InterpreterState {
             }
             "std_pad" => Box::new(combinational::StdPad::new(params, cell_qin)),
             // State components
-            "std_reg" => Box::new(stateful::StdReg::new(params, cell_qin)),
+            "std_reg" => Box::new(stateful::mem::StdReg::new(params, cell_qin)),
             "std_mem_d1" => {
                 let init = mems.as_mut().and_then(|x| x.remove(cell_name));
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::StdMemD1::from_initial_mem(
+                        Box::new(stateful::mem::StdMemD1::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::StdMemD1::new(
+                    None => Box::new(stateful::mem::StdMemD1::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -282,14 +282,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::StdMemD2::from_initial_mem(
+                        Box::new(stateful::mem::StdMemD2::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::StdMemD2::new(
+                    None => Box::new(stateful::mem::StdMemD2::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -301,14 +301,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::StdMemD3::from_initial_mem(
+                        Box::new(stateful::mem::StdMemD3::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::StdMemD3::new(
+                    None => Box::new(stateful::mem::StdMemD3::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -320,14 +320,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::StdMemD4::from_initial_mem(
+                        Box::new(stateful::mem::StdMemD4::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::StdMemD4::new(
+                    None => Box::new(stateful::mem::StdMemD4::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -339,14 +339,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::SeqMemD1::from_initial_mem(
+                        Box::new(stateful::mem::SeqMemD1::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::SeqMemD1::new(
+                    None => Box::new(stateful::mem::SeqMemD1::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -358,14 +358,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::SeqMemD2::from_initial_mem(
+                        Box::new(stateful::mem::SeqMemD2::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::SeqMemD2::new(
+                    None => Box::new(stateful::mem::SeqMemD2::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -377,14 +377,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::SeqMemD3::from_initial_mem(
+                        Box::new(stateful::mem::SeqMemD3::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::SeqMemD3::new(
+                    None => Box::new(stateful::mem::SeqMemD3::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
@@ -396,14 +396,14 @@ impl InterpreterState {
 
                 match init {
                     Some(vals) => {
-                        Box::new(stateful::SeqMemD4::from_initial_mem(
+                        Box::new(stateful::mem::SeqMemD4::from_initial_mem(
                             params,
                             cell_qin,
                             configs.allow_invalid_memory_access,
                             vals,
                         )?)
                     }
-                    None => Box::new(stateful::SeqMemD4::new(
+                    None => Box::new(stateful::mem::SeqMemD4::new(
                         params,
                         cell_qin,
                         configs.allow_invalid_memory_access,
